Raise AttributeError if inactive when `request.tm` is accessed
This makes it much harder for an application to mistakenly get hold of a
reference to the transaction manager during a request for which the
activate_hook returns `False`.

There should never be any need to get a reference to the transaction
manager during a request that has been opted out. The ability to do so
opens up possibilities for subtle bugs where database sessions are
registered with a manager that isn't itself hooked up to the request
lifecycle, resulting in leaked references to those sessions inside
`zope.sqlalchemy`.